+++ spec/js_routes/module_types/dts/routes.spec.d.ts
@@ -86,7 +86,7 @@ export const serialize: RouterExposedMethods['serialize'];
 
 /**
  * Generates rails route to
- * /inboxes/:inbox_id/messages/:message_id/attachments/:id(.:format)
+ * /inboxes/:inbox_id/messages/:message_id/attachments/:id
  * @param {any} inbox_id
  * @param {any} message_id
  * @param {any} id
@@ -97,7 +97,7 @@ export const inbox_message_attachment_path: ((
   inbox_id: RequiredRouteParameter,
   message_id: RequiredRouteParameter,
   id: RequiredRouteParameter,
-  options?: {format?: OptionalRouteParameter} & RouteOptions
+  options?: RouteOptions
 ) => string) & RouteHelperExtras;
 
 /**